Clean up - Deleted unused files and moved folders
