--- My Business Card/Controllers/CardsVCExtensions/CVC+CollectionviewDelegate.swift
@@ -1,35 +1,0 @@
-//
-//  CardsCiewController+CollectionviewDelegate.swift
-//  My Business Card
-//
-//  Created by Irving Martinez on 6/26/19.
-//  Copyright © 2019 Irving Martinez. All rights reserved.
-//
-
-import UIKit
-
-/* MARK: - CollectionView delegates and datasource methods
-extension CardsViewController: UICollectionViewDelegate, UICollectionViewDataSource {
-    
-    // Set the amount of cards
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return cards.count
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        // Create cell
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "CardCell", for: indexPath) as! CardCollectionViewCell
-        
-        // Set cell properties
-        cell.delegate = self
-        cell.card = cards[indexPath.item]
-        
-        // Return cell
-        return cell
-        
-    }
-    
-    
-    
-}
- */
