--- My Business Card/Controllers/CardsVCExtensions/CVC+CustomDelegates.swift
@@ -1,20 +1,0 @@
-//
-//  CardsViewController+CustomDelegates.swift
-//  My Business Card
-//
-//  Created by Irving Martinez on 6/26/19.
-//  Copyright © 2019 Irving Martinez. All rights reserved.
-//
-
-import UIKit
-import CoreData
-
-
-// Show error
-//extension CardsViewController: CellDelegate {
-//    func showError(error: Bool) {
-//        Alert.unableToLoad(vc: self)
-//        collectionView.reloadData()
-//    }
-//}
-
