+++ My Business Card/Protocols/Deletable.swift
@@ -1,0 +1,13 @@
+//
+//  Deletable.swift
+//  QRDS
+//
+//  Created by Irving Martinez on 1/11/20.
+//  Copyright © 2020 Irving Martinez. All rights reserved.
+//
+
+import Foundation
+
+protocol Deletable {
+    func deleteTapped(index: IndexPath)
+}
